+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Default` for `ZipSorted`
 - `Default` for `Compress`
 - `Default` for `Concurrent`
+- `Default` for `group_sequential::Grouped`
 
 ### Changed